+++ Sources/SuperwallKit/Analytics/Internal Tracking/Trackable Events/TrackableSuperwallEvent.swift
@@ -422,6 +422,7 @@ enum InternalSuperwallEvent {
       if let demandTier = demandTier {
         params["attr_demandTier"] = demandTier
       }
+      params["user_attributes"] = Superwall.shared.userAttributes
       return params
     }
     var audienceFilterParams: [String: Any] {
@@ -636,10 +637,7 @@ enum InternalSuperwallEvent {
         if let demandTier = demandTier {
           placementParams["attr_demandTier"] = demandTier
         }
-        let appleSearchAttributes = Superwall.shared.userAttributes.filter {
-          $0.key.hasPrefix("apple_search_ads_")
-        }
-        placementParams += appleSearchAttributes
+        placementParams["user_attributes"] = Superwall.shared.userAttributes
         fallthrough
       case .start,
         .abandon,