+++ Sources/SuperwallKit/Storage/Cache/CacheKeys.swift
@@ -246,5 +246,5 @@ enum IntegrationAttributes: Storable {
     "store.IntegrationAttributes"
   }
   static var directory: SearchPathDirectory = .userSpecificDocuments
-  typealias Value = [String: Any]
+  typealias Value = [String: String]
 }